requirements.txt
click
python-dotenv
huggingface_hub